+++ site/src/api/api.ts
@@ -4,7 +4,6 @@ import { MockAuditLog } from "testHelpers/entities"
 import * as Types from "./types"
 import { WorkspaceBuildTransition } from "./types"
 import * as TypesGen from "./typesGenerated"
-import { User } from "./typesGenerated"
 
 const CONTENT_TYPE_JSON: AxiosRequestHeaders = {
   "Content-Type": "application/json",
@@ -386,36 +385,9 @@ export const getEntitlements = async (): Promise<TypesGen.Entitlements> => {
   return response.data
 }
 
-export interface AuditLog {
-  readonly id: string
-  readonly request_id: string
-  readonly time: string
-  readonly organization_id: string
-  // Named type "net/netip.Addr" unknown, using "any"
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  readonly ip: any
-  readonly user_agent: string
-  readonly resource_type: "organization" | "template" | "template_version" | "user" | "workspace"
-  readonly resource_id: string
-  readonly resource_target: string
-  readonly action: any
-  readonly diff: Record<string, { old: string; new: string }>
-  readonly status_code: number
-  // This is likely an enum in an external package ("encoding/json.RawMessage")
-  readonly additional_fields: any
-  readonly description: string
-  readonly user?: User
-  // This is likely an enum in an external package ("encoding/json.RawMessage")
-  readonly resource:
-    | TypesGen.Organization
-    | TypesGen.Template
-    | TypesGen.TemplateVersion
-    | TypesGen.User
-    | TypesGen.Workspace
-}
-
-export const getAuditLogs = async (): Promise<AuditLog[]> => {
+export const getAuditLogs = async (): Promise<TypesGen.AuditLog[]> => {
   return [MockAuditLog]
+  // TODO: Uncomment this to get the data from the API instead of mock
   // const response = await axios.get("/api/v2/audit")
   // return response.data
 }
