+++ site/src/pages/AuditPage/AuditPageView.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentMeta, Story } from "@storybook/react"
-import { MockAuditLog, MockAuditLogWithDiff } from "testHelpers/entities"
+import { MockAuditLog, MockAuditLog2 } from "testHelpers/entities"
 import { AuditPageView } from "./AuditPageView"
 
 export default {
@@ -11,7 +11,7 @@ const Template: Story = (args) => <AuditPageView {...args} />
 
 export const AuditPage = Template.bind({})
 AuditPage.args = {
-  auditLogs: [MockAuditLog, MockAuditLog, MockAuditLogWithDiff],
+  auditLogs: [MockAuditLog, MockAuditLog2],
 }
 
 export const AuditPageSmallViewport = Template.bind({})
